Added duration, elapsed, shuffle, repeat implementation
Removed _xeninfo_currentNowPlayingInfo and just added
_xeninfo_MPUNowPlayingController to get elapsed and duration

Shuffle and repeat turned out to be much more difficult but was able to
get access to MPMusicPlayerController but it’s finicky. Need to verify
music is playing and still add a delay before called. It can go without
a delay but sometimes it will just freeze the device.

Needs tested on multiple devices (Tested 11.1.2 iPX)